Removed honggfuzz and wrote my own fuzzer in python.
I initially wrote it in Rust, but I ran into some issues with multithreading it (that I also ran into with python, but ended up getting around it with bash scripting).
I might re-write it in the future if it turns out that the overhead of using python is actually affecting the fuzzer's performance.